requirement.txt
Django==2.2.9
djangorestframework==3.9.2
fuzzywuzzy==0.17.0
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
numpy==1.16.4
pandas==0.24.2
PyMySQL==0.9.3
python-dateutil==2.8.0
python-Levenshtein==0.12.0
pytz==2019.1
PyYAML==5.1.1
six==1.12.0
sqlparse==0.4.2
zope.event==4.5.0
zope.interface==5.4.0
uWSGI==2.0.18
sights==1.3.8.5
